Separate thread-safe ContextManager module
Separates the sync module, which defines a thread-safe ContextManager
and was previously defined in the manager module, to a new file.

Also creates a new manager folder for the implementations of the
Context Manager to maintain the module structure.
